CCamera/CCameraScript 클래스 추가, 상수버퍼 대응구조체 전역변수로 선언(g_tTransform), View/Projection 변환 구현 중

+++ Project/Engine/CGameObject.cpp
@@ -11,7 +11,7 @@ CGameObject::CGameObject()
 CGameObject::~CGameObject()
 {
 	// basicComp
-	uint32 size = (uint32)COMPONENT_TYPE::END;
+	size_t size = (uint32)COMPONENT_TYPE::END;
 	for (size_t i = 0; i < size; i++)
 	{
 		if (nullptr != m_arrBasicComp[i])
